feat: move drivers task-executor -> algorithm-operator
