+++ core/algorithm-operator/config/main/config.base.js
@@ -15,6 +15,14 @@ config.algorithmQueueBalancer = {
     limit: formatter.parseInt(process.env.ALGORITHM_QUEUE_LIMIT, 5)
 };
 
+config.driversSetting = {
+    name: 'pipeline-driver',
+    concurrency: formatter.parseInt(process.env.PIPELINE_DRIVERS_CONCURRENCY_LIMIT, 5),
+    minAmount: formatter.parseInt(process.env.PIPELINE_DRIVERS_AMOUNT, 5),
+    scalePercent: parseFloat(process.env.PIPELINE_DRIVERS_SCALE_PERCENT || 0.5),
+    reconcileInterval: formatter.parseInt(process.env.PIPELINE_DRIVERS_RECONCILE_INTERVAL, 5000)
+};
+
 config.db = {
     provider: 'mongo',
     mongo: {
