+++ core/algorithm-operator/config/main/config.test.js
@@ -1,3 +1,7 @@
 const config = {};
-
+config.driversSetting = {
+    minAmount: 20,
+    scalePercent: 0.2,
+    concurrency: 5,
+};
 module.exports = config;
