+++ core/algorithm-operator/lib/consts/commands.js
@@ -1,0 +1,3 @@
+module.exports = {
+    stopProcessing: 'stopProcessing'
+};
