+++ core/algorithm-operator/lib/consts/componentNames.js
@@ -7,7 +7,8 @@ const Components = {
     JOBS_PRODUCER: 'JobProducer',
     ALGORITHM_BUILDS_RECONCILER: 'Algorithm-Builds-Reconciler',
     ALGORITHM_DEBUG_RECONCILER: 'Algorithm-Debug-Reconciler',
-    ALGORITHM_QUEUE_RECONCILER: 'Algorithm-Queue-Reconciler'
+    ALGORITHM_QUEUE_RECONCILER: 'Algorithm-Queue-Reconciler',
+    DRIVERS_RECONCILER: 'Drivers-Reconciler'
 };
 
 module.exports = Components;
