+++ core/algorithm-operator/lib/helpers/utils.js
@@ -1,0 +1,11 @@
+const arrayToMap = (array) => {
+    const init = Object.create(null);
+    return array.reduce((map, obj) => {
+        map[obj.name] = obj;  // eslint-disable-line
+        return map;
+    }, init);
+};
+
+module.exports = {
+    arrayToMap,
+};
