+++ core/algorithm-operator/lib/reconcile/normalize.js
@@ -118,6 +118,54 @@ const normalizeSecret = (secret) => {
     return secret.body;
 };
 
+const normalizeDrivers = (drivers) => {
+    if (!drivers) {
+        return [];
+    }
+    const driversArray = drivers.map((d) => {
+        return {
+            id: d.driverId,
+            idle: d.idle,
+            paused: d.paused,
+            podName: d.podName,
+            jobs: d.jobs?.length || 0
+        };
+    });
+    return driversArray;
+};
+
+const normalizeDriversRequests = (requests, name) => {
+    if (requests == null || requests.length === 0 || requests[0].data == null) {
+        return 0;
+    }
+    return requests[0].data.filter(r => r.name === name).length;
+};
+
+const normalizeDriversJobs = (jobsRaw, predicate = () => true) => {
+    if (!jobsRaw || !jobsRaw.body || !jobsRaw.body.items) {
+        return [];
+    }
+    const jobs = jobsRaw.body.items
+        .filter(predicate)
+        .map(j => ({
+            name: j.metadata.name,
+            active: j.status.active === 1
+        }));
+    return jobs;
+};
+
+const normalizeDriversAmount = (drivers, requests, settings) => {
+    const { minAmount, maxAmount, concurrency } = settings;
+    const available = drivers.map(d => concurrency - d.jobs).reduce((a, b) => a + b, 0);
+    let amount = minAmount;
+
+    if (requests > available) {
+        amount = (requests - available) / concurrency;
+    }
+    const desiredDrivers = Math.min(amount, maxAmount);
+    return desiredDrivers;
+};
+
 module.exports = {
     normalizeQueuesDeployments,
     normalizeDebugDeployments,
@@ -126,5 +174,9 @@ module.exports = {
     normalizeAlgorithms,
     normalizeBuildJobs,
     normalizeBoardDeployments,
-    normalizeSecret
+    normalizeSecret,
+    normalizeDrivers,
+    normalizeDriversRequests,
+    normalizeDriversJobs,
+    normalizeDriversAmount
 };
