+++ core/algorithm-operator/lib/templates/pipeline-driver.js
