+++ core/algorithm-operator/tests/stub/driversTemplateStore.json
