+++ core/task-executor/config/main/config.test.js
@@ -1,9 +1,5 @@
 const config = {};
-config.driversSetting = {
-    minAmount: 20,
-    scalePercent: 0.2,
-    concurrency: 5,
-};
+
 config.intervalMs = process.env.INTERVAL_MS || 15000;
 
 const kubeconfig = {
