+++ core/task-executor/lib/consts/containers.js
@@ -1,5 +1,4 @@
 module.exports = {
-    PIPELINE_DRIVER: 'pipeline-driver',
     WORKER: 'worker',
     ALGORITHM: 'algorunner'
 };
