+++ core/task-executor/lib/templates/index.js
@@ -1,7 +1,3 @@
 const workerTemplate = require('./worker');
-const pipelineDriverTemplate = require('./pipeline-driver');
 
-module.exports = {
-    ...workerTemplate,
-    pipelineDriverTemplate
-};
+module.exports = workerTemplate;
